+++ scripts/start-root-ca.sh
@@ -6,20 +6,21 @@
 #
 
 set -ex
+if [ ! -f $TARGET_CERTFILE ]; then
+  # Initialize the root CA
+  fabric-ca-server init -b $BOOTSTRAP_USER_PASS
 
-# Initialize the root CA
-fabric-ca-server init -b $BOOTSTRAP_USER_PASS
+  # Copy the root CA's signing certificate to the data directory to be used by others
+  cp $FABRIC_CA_SERVER_HOME/ca-cert.pem $TARGET_CERTFILE
 
-# Copy the root CA's signing certificate to the data directory to be used by others
-cp $FABRIC_CA_SERVER_HOME/ca-cert.pem $TARGET_CERTFILE
-
-# Add the custom orgs
-for o in $FABRIC_ORGS; do
-   aff=$aff"\n   $o: []"
-done
-aff="${aff#\\n   }"
-sed -i "/affiliations:/a \\   $aff" \
-   $FABRIC_CA_SERVER_HOME/fabric-ca-server-config.yaml
+  # Add the custom orgs
+  for o in $FABRIC_ORGS; do
+     aff=$aff"\n   $o: []"
+  done
+  aff="${aff#\\n   }"
+  sed -i "/affiliations:/a \\   $aff" \
+     $FABRIC_CA_SERVER_HOME/fabric-ca-server-config.yaml
+fi
 
 # Start the root CA
 fabric-ca-server start
